내 정보 가져오기 및 순환참조 문제 해결
- 명세서 기준으로 변경하였습니다.
- 인자로 Entity를 넘겨 순환참조가 되던 문제를 해결했습니다.

+++ ASSA_backend/assa-backend/src/main/kotlin/com/assa/assabackend/entity/RegionDistrict.kt
@@ -11,10 +11,10 @@ data class RegionDistrict (
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "region_district_id")
-    val regionDistrictId: Int,
+    val regionDistrictId: Long,
 
     @Column(name="region_metro_id", nullable = false)
-    val regionMetroId: Int,
+    val regionMetroId: Long,
 
     @Column(name="name", length = 50, nullable = false)
     val name: String,
